// src/models/digital-agency/brand.user.ts
export class BrandUsers {
    id: number;
    accountVerified: boolean;
    created: number;
    firstName: string;
    lastName: string;
    gender: string;
    lastLogin: any;
    loginEmail: string;
    profilePic: any;
    token: any;
    updated: any;
    userRole: {
        id: number;
        roleName: string;
    };
    welcomeMessage: boolean;
}